refactor: switch 'game' to use 'AbstractView' class
