+++ js/game/game.js
@@ -1,18 +1,19 @@
 // Game controller window
 
+import GameView from './game-view.js';
 import GameData from './game-data.js';
-import gameWindow from './game-window.js';
 
 import {getAnswers} from './game-utils.js';
 import getQuestions from '../data/mock-questions.js';
 import goBeginWindow from '../greeting/greeting.js';
 import goStatistics from '../stats/stats-window';
-import {isAnsweredYes} from '../utils.js';
+import {changeWindow, getCountInputsChecked, isAnsweredYes} from '../utils.js';
 
 const QUESTIONS_IS_NOT_SHAFFLED = true;
 const CONTINUE_STARTED_GAME = true;
 const MILLISECONDS_TICK = 1000;
 const GAME_CANCELLING_CONFIRMATION_TEXT = `Вы уверены что хотите начать игру заново и сбросить результаты?`;
+const INPUTS_SELECTOR = `.game__content .game__option input`;
 
 const gameQuestions = getQuestions(QUESTIONS_IS_NOT_SHAFFLED);
 const gameData = GameData;
@@ -37,12 +38,29 @@ const processAnswer = (evt) => {
   }
 };
 
+const verifyUserAnswerClick = (evt) => {
+  const inputNodes = [...document.querySelectorAll(INPUTS_SELECTOR)];
+  if (
+    (inputNodes.length && isAllInputsSelected(inputNodes)) ||
+    !inputNodes.length) {
+    processAnswer(evt);
+  }
+};
+
+const isAllInputsSelected = (inputNodes) => {
+  return getCountInputsChecked(inputNodes) >= gameData.currentQuestionImageCount;
+};
+
 const run = (isToBeContinued) => {
   if (!isToBeContinued) {
     gameData.init(gameQuestions, processAnswer);
   }
   timerID = setInterval(gameData.tickSecond.bind(gameData), MILLISECONDS_TICK);
-  gameWindow(gameData, confirmCancellingGame, processAnswer);
+
+  const thisWindow = new GameView(gameData);
+  thisWindow.onAnswerClick = verifyUserAnswerClick;
+  thisWindow.onCancelGameClick = confirmCancellingGame;
+  changeWindow(thisWindow.element);
 };
 
 export default run;
